Cambios en Rest-User y Rest-Service

+++ Rest-Service/src/main/java/com/rest/service/repository/ReclamoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rest.service.model.Reclamo;
 
-public interface ReclamoRepository extends JpaRepository<Reclamo, Long> {
+public interface ReclamoRepository extends JpaRepository<Reclamo, Integer> {
 
 }
